requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
werkzeug>=3.0.0
joblib>=1.3.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
